Move UI files to a UI directory and adjusted file paths for data resources.
