+++ src/ui/character_picker_dialog.py
@@ -1,9 +1,9 @@
 """suggestion_dialog.py."""
 
-from src.dialog import Dialog
+from src.ui.dialog import Dialog
 from src.card import Card
 from src.cardtype import CardType
-from src.character_picker import CharacterPicker
+from src.ui.character_picker import CharacterPicker
 from src.network.message import Message
 from src.network.message import MessageType
 from src.network.clientnetworkinterface import ClientNetworkInterface as CNI
@@ -16,7 +16,7 @@ class CharacterPickerDialog(Dialog):
         super(CharacterPickerDialog, self).__init__(x, y)
         # Initialize pickers
         dir = os.path.dirname(__file__)
-        filename = os.path.join(dir, '../data/cards.json')
+        filename = os.path.join(dir, '../../data/cards.json')
 
         with open(filename) as data_file:
             card_data = json.load(data_file)
